RISDEV-12577 Improve tests
* Split html endpoint tests out of NormsControllerApiIntegrationTest
